+++ Changelog.md
@@ -9,6 +9,10 @@ every new version is a new major version.
 
 ## 109 - 2023-09-21
 
+### Changed
+
+-   Improve `Nrfutil Sandbox` error messages.
+
 ### Added
 
 -   `Overlay` component.